files are added...

+++ shell-scripting/roboshop/common.sh
@@ -1,5 +1,11 @@
 #!/usr/bin/bash
 
+USER_ID=$(id -u)
+if [ $USER-ID -ne 0 ]; then
+  echo -e " You should be the root user / sudo user too run this script"
+  exit 2
+fi
+
 LOG=/tmp/roboshop.log
 rm -f $LOG
 
@@ -8,10 +14,12 @@ STAT_CHECK() {
   echo done
 else
   echo fail
+  echo -e " Check the log file more details, Log File - $LOG"
   exit 1
 fi
 }
 
 PRINT() {
-  echo -n -e " $1\t\t..."
+  echo -e "#################\t$1###################" &>>
+  echo -n -e "$1\t\t..."
 }
